Refactor user retention validation as composable
